app/pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mascan"
version = "2.0.0"
description = "MaScan - QR Code Attendance System"
readme = "FLASK_README.md"
requires-python = ">=3.9"
authors = [
    { name = "Group 12", email = "group12@example.com" }
]
license = { text = "MIT" }

dependencies = [
    "flask==3.0.0",
    "flask-cors==4.0.0",
    "flask-session==0.5.0",
    "python-dotenv==1.0.0",
    "bcrypt==4.1.2",
    "werkzeug==3.0.1",
    "reportlab==4.0.7",
    "Pillow==10.1.0",
    "python-qrcode==7.4.2",
    "numpy==1.24.4",
    "python-dateutil==2.8.2"
]